+++ eslint.config.mjs
@@ -1,13 +1,11 @@
 // @ts-check
 
 import { defineConfig } from 'eslint/config';
-import eslint from '@eslint/js';
 import globals from 'globals';
 import appiumConfig from '@appium/eslint-config-appium-ts';
 
 
 export default defineConfig(
-  eslint.configs.recommended,
   ...appiumConfig,
   {
     files: ['test/e2e/**/*.ts'],